+++ CHANGELOG.md
@@ -1,5 +1,3 @@
 # Changelog
 
-* Updates Account to have an accountType field.
-* Updates GetAccounts to return accounts with the new accountType field.
-* Updates getLoginRoles to return the accountType field.
+* Updates GetIAMKeys to include the optional Category field.